@@ -1,10 +1,10 @@
 package org.systemexception.simplexdb.test;
 
-import org.junit.Before;
-import org.junit.Test;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 import org.systemexception.simplexdb.domain.Data;
 
-import static org.junit.Assert.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 /**
  * @author leo
@@ -15,7 +15,7 @@ public class DataTest {
 	private Data sut;
 	private final byte[] bytes = new byte[256];
 
-	@Before
+	@BeforeEach
 	public void setUp() {
 		String dataName = "TEST_DATA";
 		sut = new Data(dataName, bytes);
@@ -23,6 +23,6 @@ public class DataTest {
 
 	@Test
 	public void data_size_is_correct() {
-		assertTrue(256L == sut.getSize());
+        assertEquals(256L, (long) sut.getSize());
 	}
 }
