@@ -1,14 +1,14 @@
 package org.systemexception.simplexdb.test.controller;
 
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.HttpStatus;
 import org.springframework.mock.web.MockMultipartFile;
 import org.springframework.test.context.TestPropertySource;
-import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import org.systemexception.simplexdb.Application;
@@ -18,10 +18,11 @@ import org.systemexception.simplexdb.database.DatabaseApi;
 import org.systemexception.simplexdb.domain.Data;
 import org.systemexception.simplexdb.service.StorageService;
 
+import java.io.File;
 import java.util.Optional;
 import java.util.UUID;
 
-import static org.mockito.Matchers.any;
+import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
@@ -30,24 +31,27 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
  * @author leo
  * @date 28/02/16 17:22
  */
-@RunWith(SpringRunner.class)
+@ExtendWith(SpringExtension.class)
 @SpringBootTest(classes = {Application.class})
 @TestPropertySource(locations = "classpath:application.properties")
 public abstract class AbstractControllerTest {
 
-	protected static String TEST_DATABASE_FULLPATH;
+	public static String TEST_DATABASE_FULLPATH;
 	protected DatabaseApi databaseService;
 	protected StorageService storageService;
 	@InjectMocks
 	@Autowired
 	protected SimplexDbController simplexDbController;
 	protected MockMvc sut;
-	private final static String ENDPOINT = Endpoints.CONTEXT, REQUEST_PARAM = Endpoints.FILE_TO_UPLOAD,
-			FILE_TEXT_FORMAT = "text/plain", FILE_TEXT_DATA = "some data in the file", URL_SEPARATOR = "/";
+	private final static String ENDPOINT = Endpoints.CONTEXT;
+    private final static String REQUEST_PARAM = Endpoints.FILE_TO_UPLOAD;
+    private final static String FILE_TEXT_FORMAT = "text/plain";
+    private final static String FILE_TEXT_DATA = "some data in the file";
+    private final static String  URL_SEPARATOR = File.separator;
 	protected Data testData;
 
 	@Test
-	public void save() throws Exception {
+	void save() throws Exception {
 		MockMultipartFile dataFile = new MockMultipartFile(REQUEST_PARAM, UUID.randomUUID().toString(),
 				FILE_TEXT_FORMAT, FILE_TEXT_DATA.getBytes());
 		sut.perform(MockMvcRequestBuilders.fileUpload(ENDPOINT + Endpoints.SAVE).file(dataFile))
@@ -58,7 +62,7 @@ public abstract class AbstractControllerTest {
 	}
 
 	@Test
-	public void save_conflict() throws Exception {
+	void save_conflict() throws Exception {
 		when(databaseService.save(any())).thenReturn(false);
 		MockMultipartFile dataFile = new MockMultipartFile(REQUEST_PARAM, UUID.randomUUID().toString(),
 				FILE_TEXT_FORMAT, FILE_TEXT_DATA.getBytes());
@@ -70,28 +74,28 @@ public abstract class AbstractControllerTest {
 	}
 
 	@Test
-	public void find_all() throws Exception {
+	void find_all() throws Exception {
 		sut.perform(MockMvcRequestBuilders.get(ENDPOINT + Endpoints.FIND)).andExpect(status()
 				.is(HttpStatus.OK.value()));
 		verify(databaseService).findAll();
 	}
 
 	@Test
-	public void view_all() throws Exception {
+	void view_all() throws Exception {
 		sut.perform(MockMvcRequestBuilders.get(ENDPOINT + Endpoints.VIEW)).andExpect(status()
 				.is(HttpStatus.OK.value()));
 		verify(databaseService).findAll();
 	}
 
 	@Test
-	public void find_id_and_save() throws Exception {
+	void find_id_and_save() throws Exception {
 		sut.perform(MockMvcRequestBuilders.get(ENDPOINT + Endpoints.FIND + URL_SEPARATOR + testData.getInternalId()))
 				.andExpect(status().is(HttpStatus.OK.value()));
 		verify(databaseService).findById(testData.getName());
 	}
 
 	@Test
-	public void dont_find_id() throws Exception {
+	void dont_find_id() throws Exception {
 		when(databaseService.findById(testData.getName())).thenReturn(Optional.empty());
 		sut.perform(MockMvcRequestBuilders.get(ENDPOINT + Endpoints.FIND + URL_SEPARATOR + testData.getInternalId()))
 				.andExpect(status().is(HttpStatus.NOT_FOUND.value()));
@@ -99,20 +103,20 @@ public abstract class AbstractControllerTest {
 	}
 
 	@Test
-	public void find_match() throws Exception {
+	void find_match() throws Exception {
 		sut.perform(MockMvcRequestBuilders.get(ENDPOINT + Endpoints.FINDBYNAME + URL_SEPARATOR + any()));
 		verify(databaseService).findByFilename(any());
 	}
 
 	@Test
-	public void delete_existing() throws Exception {
+	void delete_existing() throws Exception {
 		sut.perform(MockMvcRequestBuilders.delete(ENDPOINT + Endpoints.DELETE + URL_SEPARATOR + testData.getName()))
 				.andExpect(status().is(HttpStatus.OK.value()));
 		verify(databaseService).delete(testData.getName());
 	}
 
 	@Test
-	public void delete_not_existing() throws Exception {
+	void delete_not_existing() throws Exception {
 		when(databaseService.delete(testData.getName())).thenReturn(false);
 		sut.perform(MockMvcRequestBuilders.delete(ENDPOINT + Endpoints.DELETE + URL_SEPARATOR + testData.getName()))
 				.andExpect(status().is(HttpStatus.NOT_FOUND.value()));
