@@ -1,7 +1,7 @@
 package org.systemexception.simplexdb.test.controller;
 
 import com.sleepycat.je.DatabaseException;
-import org.junit.Before;
+import org.junit.jupiter.api.BeforeEach;
 import org.mockito.MockitoAnnotations;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 import org.systemexception.simplexdb.controller.SimplexDbController;
@@ -11,6 +11,7 @@ import org.systemexception.simplexdb.service.StorageService;
 import org.systemexception.simplexdb.test.database.AbstractDbTest;
 import org.systemexception.simplexdb.test.database.MapDbServiceTest;
 
+import java.io.File;
 import java.io.IOException;
 import java.util.Optional;
 
@@ -20,11 +21,11 @@ import static org.mockito.Mockito.*;
  * @author leo
  * @date 05/12/15 21:53
  */
-public class MapDbTest extends AbstractControllerTest {
+class MapDbTest extends AbstractControllerTest {
 
-	@Before
-	public void setUp() throws DatabaseException, IOException, ClassNotFoundException {
-		TEST_DATABASE_FULLPATH = AbstractDbTest.TARGET_FOLDER + "/" +
+	@BeforeEach
+	void setUp() throws DatabaseException, IOException, ClassNotFoundException {
+		TEST_DATABASE_FULLPATH = AbstractDbTest.TARGET_FOLDER + File.separator +
 				MapDbServiceTest.TEST_DATABASE_FILENAME;
 		testData = new Data();
 		testData.setInternalId("123");
