@@ -2,13 +2,14 @@ package org.systemexception.simplexdb.test.database;
 
 import com.sleepycat.je.DatabaseException;
 import org.apache.commons.io.FileUtils;
-import org.junit.After;
-import org.junit.AfterClass;
-import org.junit.Test;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.Test;
 import org.systemexception.simplexdb.database.DatabaseApi;
+import org.systemexception.simplexdb.database.impl.BerkeleyDbService;
 import org.systemexception.simplexdb.domain.Data;
 import org.systemexception.simplexdb.service.StorageService;
 import org.systemexception.simplexdb.service.StorageServiceApi;
+import org.systemexception.simplexdb.test.controller.AbstractControllerTest;
 
 import java.io.File;
 import java.io.IOException;
@@ -20,7 +21,7 @@ import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.stream.Stream;
 
-import static org.junit.Assert.*;
+import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.mock;
 
 /**
@@ -29,34 +30,40 @@ import static org.mockito.Mockito.mock;
  */
 public abstract class AbstractDbTest {
 
-	protected DatabaseApi sut;
-	protected static String TEST_DATABASE_FULLPATH;
-	protected final StorageServiceApi storageServiceApi = mock(StorageService.class);
-	public static final String TARGET_FOLDER = "target", TEST_DATABASE_ID = "dataId";
+    protected DatabaseApi sut;
 
-	@After
-	public void tearDown() throws IOException {
-		sut.close();
-		File databaseFile = new File(TEST_DATABASE_FULLPATH);
-		Stream<Path> walk = Files.walk(Paths.get(TEST_DATABASE_FULLPATH), FileVisitOption.FOLLOW_LINKS);
-		walk.forEach(item -> item.toFile().delete());
-		FileUtils.deleteDirectory(new File(TEST_DATABASE_FULLPATH));
+    public static final String TARGET_FOLDER = "target";
+    public static final String TEST_DATABASE_ID = "dataId";
 
-		assertFalse(databaseFile.exists());
-	}
+    protected static final StorageServiceApi STORAGE_SERVICE_API = mock(StorageService.class);
 
-	@AfterClass
-	public static void destroySut() throws IOException {
-		File databaseFile = new File(TEST_DATABASE_FULLPATH);
-		if (databaseFile.exists()) {
-			Stream<Path> walk = Files.walk(Paths.get(TEST_DATABASE_FULLPATH), FileVisitOption.FOLLOW_LINKS);
-			walk.forEach(item -> item.toFile().delete());
-			databaseFile.deleteOnExit();
-		}
-	}
+	@AfterEach
+	void tearDown() throws IOException {
+        if (sut instanceof BerkeleyDbService) {
+            sut.close();
+        }
+        File databaseFile = new File(AbstractControllerTest.TEST_DATABASE_FULLPATH);
+        Stream<Path> walk = Files.walk(Paths.get(AbstractControllerTest.TEST_DATABASE_FULLPATH), FileVisitOption.FOLLOW_LINKS);
+        walk.forEach(item -> item.toFile().delete());
+        FileUtils.deleteDirectory(new File(AbstractControllerTest.TEST_DATABASE_FULLPATH));
+        assertFalse(databaseFile.exists());
+    }
+
+    /*
+    @AfterAll
+    public static void destroySut() throws IOException {
+        File databaseFile = new File(TEST_DATABASE_FULLPATH);
+        if (databaseFile.exists()) {
+            Stream<Path> walk = Files.walk(Paths.get(TEST_DATABASE_FULLPATH), FileVisitOption.FOLLOW_LINKS);
+            walk.forEach(item -> item.toFile().delete());
+            databaseFile.deleteOnExit();
+        }
+    }
+
+     */
 
 	@Test
-	public void add_record() throws DatabaseException, IOException {
+	void add_record() throws DatabaseException, IOException {
 		Data data = getDataForDatabase("id");
 		boolean saved = sut.save(data);
 
@@ -64,7 +71,7 @@ public abstract class AbstractDbTest {
 	}
 
 	@Test
-	public void dont_add_duplicate_record() throws DatabaseException, IOException {
+	void dont_add_duplicate_record() throws DatabaseException, IOException {
 		Data data = getDataForDatabase("id");
 		boolean saved = sut.save(data);
 		assertTrue(saved);
@@ -74,32 +81,32 @@ public abstract class AbstractDbTest {
 	}
 
 	@Test
-	public void get_data_id_list() throws DatabaseException, IOException, ClassNotFoundException {
+	void get_data_id_list() throws DatabaseException, IOException, ClassNotFoundException {
 		int dataToAdd = 10;
 		for (int i = 0; i < dataToAdd; i++) {
 			Data data = getDataForDatabase(String.valueOf(i));
 			sut.save(data);
 		}
-		assertTrue(sut.findAll().size() == dataToAdd);
+        assertEquals(sut.findAll().size(), dataToAdd);
 	}
 
 	@Test
-	public void delete_existing_data() throws DatabaseException, IOException {
+	void delete_existing_data() throws DatabaseException, IOException {
 		Data data = getDataForDatabase("id");
 		sut.save(data);
 
-		assertTrue(sut.delete(data.getInternalId()));
+		assertTrue(sut.delete(data.getName()));
 	}
 
 	@Test
-	public void dont_delete_non_existing_data() throws DatabaseException {
-		Data data = getDataForDatabase("id");
+	void dont_delete_non_existing_data() throws DatabaseException {
+		Data data = new Data("internalId", "someName", System.currentTimeMillis(), "SomeData".getBytes());
 
 		assertFalse(sut.delete(data.getName()));
 	}
 
 	@Test
-	public void find_existing_data() throws DatabaseException, IOException, ClassNotFoundException {
+	void find_existing_data() throws DatabaseException, IOException, ClassNotFoundException {
 		Data data = getDataForDatabase("id");
 		sut.save(data);
 		Data foundData = sut.findById(data.getInternalId()).get();
@@ -108,16 +115,16 @@ public abstract class AbstractDbTest {
 		assertEquals(data.getSize(), foundData.getSize());
 	}
 
-	@Test(expected = NoSuchElementException.class)
-	public void dont_find_non_existing_data() throws DatabaseException, IOException, ClassNotFoundException {
+	@Test
+	void dont_find_non_existing_data() throws DatabaseException, IOException, ClassNotFoundException {
 		String nonExistingId = "nonExistingId";
-		Data emptyData = sut.findById(nonExistingId).get();
-
-		assertTrue(null == emptyData);
+        assertThrows(NoSuchElementException.class, () -> {
+            Data emptyData = sut.findById(nonExistingId).get();
+        });
 	}
 
 	@Test
-	public void find_matches() throws DatabaseException, IOException, ClassNotFoundException {
+	void find_matches() throws DatabaseException, IOException, ClassNotFoundException {
 		int dataToAdd = 5;
 		for (int i = 0; i < dataToAdd; i++) {
 			Data data = getDataForDatabase(String.valueOf(i));
@@ -125,17 +132,17 @@ public abstract class AbstractDbTest {
 		}
 
 		List<Data> foundItems = sut.findByFilename("1");
-		assertTrue(foundItems.size() == 1);
+        assertEquals(1, foundItems.size());
 
 		foundItems = sut.findByFilename("data");
-		assertTrue(foundItems.size() == 5);
+        assertEquals(5, foundItems.size());
 
 		foundItems = sut.findByFilename("NON_EXISTING_ID");
-		assertTrue(foundItems.size() == 0);
+        assertEquals(0, foundItems.size());
 	}
 
 	@Test
-	public void data_integrity() throws IOException, ClassNotFoundException {
+	void data_integrity() throws IOException, ClassNotFoundException {
 		Data data = new Data("123","dataName",123456L,"dataContent".getBytes());
 		sut.save(data);
 		Data dataFech = sut.findById("123").get();
