@@ -1,19 +1,20 @@
 package org.systemexception.simplexdb.test.database;
 
-import org.junit.After;
-import org.junit.AfterClass;
-import org.junit.Before;
-import org.junit.Test;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 import org.systemexception.simplexdb.database.AbstractDbService;
 import org.systemexception.simplexdb.database.DatabaseApi;
 import org.systemexception.simplexdb.database.impl.MapDbService;
 import org.systemexception.simplexdb.domain.Data;
+import org.systemexception.simplexdb.test.controller.AbstractControllerTest;
 
 import java.io.File;
 import java.io.IOException;
 import java.util.List;
 
-import static org.junit.Assert.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
 
 /**
  * @author leo
@@ -21,52 +22,43 @@ import static org.junit.Assert.assertTrue;
  */
 public class MapDbServiceTest extends AbstractDbTest {
 
-	public static final String TEST_DATABASE_FILENAME = "test_map.db";
+	public static final String TEST_DATABASE_FILENAME = "target/test_map.db";
 
-	@Before
+	@BeforeEach
 	public void setUp() {
-		TEST_DATABASE_FULLPATH = AbstractDbTest.TARGET_FOLDER + "/" + TEST_DATABASE_FILENAME;
+		AbstractControllerTest.TEST_DATABASE_FULLPATH = AbstractDbTest.TARGET_FOLDER + File.separator + TEST_DATABASE_FILENAME;
 		File databaseFile = new File(TEST_DATABASE_FILENAME);
 		if (databaseFile.exists()) {
 			databaseFile.delete();
 		}
-		sut = new MapDbService(storageServiceApi, TEST_DATABASE_FILENAME, 1000L);
+		sut = new MapDbService(STORAGE_SERVICE_API, TEST_DATABASE_FILENAME, 1000L);
 	}
 
-	@AfterClass
-	public static void tearDownSut() {
-		File databaseFile = new File(TEST_DATABASE_FILENAME);
-		if (databaseFile.exists()) {
-			databaseFile.deleteOnExit();
-		}
-	}
+    @AfterEach
+    public void tearDown() {
+        sut.close();
+    }
 
 	@Test
-	public void limit_memory() throws IOException, ClassNotFoundException {
+	void limit_memory() throws IOException, ClassNotFoundException {
 		DatabaseApi innerSut;
-		innerSut = new MapDbService(storageServiceApi, AbstractDbTest.TARGET_FOLDER + "/" +
+		innerSut = new MapDbService(STORAGE_SERVICE_API, AbstractDbTest.TARGET_FOLDER + File.separator +
 				"low_mem_mapdb_test_db_1", 1L);
 		innerSut.save(getDataForDatabase(AbstractDbTest.TEST_DATABASE_ID));
 		List<Data> dataId = innerSut.findByFilename(AbstractDbTest.TEST_DATABASE_ID);
 
-		assertTrue(dataId.size() == 1);
-		assertTrue(AbstractDbService.WARNING_MESSAGE_MEMORY_OCCUPATION.equals(dataId.get(0).getInternalId()));
+        assertEquals(1, dataId.size());
+        assertEquals(AbstractDbService.WARNING_MESSAGE_MEMORY_OCCUPATION, dataId.get(0).getInternalId());
 	}
 
 	@Test
-	public void limit_memory_findall() throws IOException, ClassNotFoundException {
+	void limit_memory_findall() throws IOException, ClassNotFoundException {
 		DatabaseApi innerSut;
-		innerSut = new MapDbService(storageServiceApi, AbstractDbTest.TARGET_FOLDER + "/" +
+		innerSut = new MapDbService(STORAGE_SERVICE_API, AbstractDbTest.TARGET_FOLDER + File.separator +
 				"low_mem_mapdb_test_db_2", 1L);
 		innerSut.save(getDataForDatabase(AbstractDbTest.TEST_DATABASE_ID));
 		List<Data> dataId = innerSut.findAll();
 
-		assertTrue(dataId.size() == 1);
+        assertEquals(1, dataId.size());
 	}
-
-	@After
-	public void tearDown() {
-		sut.close();
-	}
-
 }
